+++ client/src/hooks/store/useLeaderBoardStore.tsx
@@ -7,10 +7,9 @@ import { useGetRealms, useRealm } from "../helpers/useRealm";
 import { useEffect, useRef, useState } from "react";
 import { useEntities } from "../helpers/useEntities";
 
-
 import { Subscription } from "rxjs";
 import useBlockchainStore from "../store/useBlockchainStore";
-import { useContributions } from "../helpers/useContributions"
+import { useContributions } from "../helpers/useContributions";
 import { ResourcesIds } from "@bibliothecadao/eternum";
 
 // Define elsewhere
@@ -18,14 +17,14 @@ const PTS_PER_CYCLE = 100;
 const CYCLE_TIME = 24 * 60 * 60 * 1000;
 
 // Computed based on qty of 10 of each resource for hyperstructure building
-const TOTAL_EFFECTIVE_CONTRIBUTION = 8209.10;
+const TOTAL_EFFECTIVE_CONTRIBUTION = 8209.1;
 
 // fix ?
-const ResourceMultipliers: { [key in ResourcesIds]?: number } = {
-  [ResourcesIds.Wood]: 1.00,
+export const ResourceMultipliers: { [key in ResourcesIds]?: number } = {
+  [ResourcesIds.Wood]: 1.0,
   [ResourcesIds.Stone]: 1.27,
   [ResourcesIds.Coal]: 1.31,
-  [ResourcesIds.Copper]: 1.90,
+  [ResourcesIds.Copper]: 1.9,
   [ResourcesIds.Obsidian]: 2.26,
   [ResourcesIds.Silver]: 2.88,
   [ResourcesIds.Ironwood]: 4.25,
@@ -33,7 +32,7 @@ const ResourceMultipliers: { [key in ResourcesIds]?: number } = {
   [ResourcesIds.Gold]: 5.49,
   [ResourcesIds.Hartwood]: 8.44,
   [ResourcesIds.Diamonds]: 16.72,
-  [ResourcesIds.Sapphire]: 20.30,
+  [ResourcesIds.Sapphire]: 20.3,
   [ResourcesIds.Ruby]: 20.98,
   [ResourcesIds.DeepCrystal]: 20.98,
   [ResourcesIds.Ignium]: 29.15,
@@ -41,7 +40,7 @@ const ResourceMultipliers: { [key in ResourcesIds]?: number } = {
   [ResourcesIds.TrueIce]: 36.06,
   [ResourcesIds.TwilightQuartz]: 45.18,
   [ResourcesIds.AlchemicalSilver]: 53.92,
-  [ResourcesIds.Adamantine]: 91.20,
+  [ResourcesIds.Adamantine]: 91.2,
   [ResourcesIds.Mithral]: 135.53,
   [ResourcesIds.Dragonhide]: 217.92,
   [ResourcesIds.Earthenshard]: 20.98,
@@ -59,7 +58,7 @@ function getResourceMultiplier(resourceType: BigInt): number {
 
 function computeContributionPoints(totalPoints: number, qty: number, type: BigInt) {
   const effectiveContribution = qty * getResourceMultiplier(type);
-  const points = effectiveContribution / TOTAL_EFFECTIVE_CONTRIBUTION * totalPoints;
+  const points = (effectiveContribution / TOTAL_EFFECTIVE_CONTRIBUTION) * totalPoints;
   return points;
 }
 
@@ -128,10 +127,8 @@ export const useComputePointsLeaderboards = () => {
   // Subscribe outside of leaderboard ?
   useEffect(() => {
     const subscribeToHyperstructureFinished = async () => {
-
       const observable = await hyperstructureFinishedEvents();
       const subscription = observable.subscribe((event) => {
-
         // Handles undefined nextBlockTimestamp
         if (event && nextBlockTimestamp) {
           const updatedPointsLeaderboard = updatePointsLeaderboard(event);
@@ -145,33 +142,35 @@ export const useComputePointsLeaderboards = () => {
         isComponentMounted.current = false;
         subscriptionRef.current?.unsubscribe();
       };
-  }}, []);
+    };
+  }, []);
 
-  const updatePointsLeaderboard = ((event: any): PlayerPointsLeaderboardInterface[] => {
+  const updatePointsLeaderboard = (event: any): PlayerPointsLeaderboardInterface[] => {
     const hyperstructureEntityId = BigInt(event.keys[0]);
     const creationTimestamp = BigInt(event.keys[1]);
-  
-    const nbOfCycles = Math.floor( Number(BigInt(nextBlockTimestamp!) - creationTimestamp) / CYCLE_TIME);
+
+    const nbOfCycles = Math.floor(Number(BigInt(nextBlockTimestamp!) - creationTimestamp) / CYCLE_TIME);
     let totalHyperstructurePoints = PTS_PER_CYCLE * nbOfCycles;
 
     // Rename temp var
     let tempPlayerPointsLeaderboard = playerPointsLeaderboard;
-    
-    const { contributionsToHyperstructure } = useContributions(hyperstructureEntityId)
-    
+
+    const { contributionsToHyperstructure } = useContributions(hyperstructureEntityId);
+
     contributionsToHyperstructure().forEach((contribution) => {
-      
       // Probably add 0x and convert to string, needs to be tested
       const playerAddress = "0x" + String(contribution.player_address);
-      
-      const index = tempPlayerPointsLeaderboard.findIndex(player => player.address === playerAddress);
-      tempPlayerPointsLeaderboard[index].totalPoints += computeContributionPoints(totalHyperstructurePoints, contribution.amount, BigInt(contribution.resource_type));
-    })
 
-    
+      const index = tempPlayerPointsLeaderboard.findIndex((player) => player.address === playerAddress);
+      tempPlayerPointsLeaderboard[index].totalPoints += computeContributionPoints(
+        totalHyperstructurePoints,
+        contribution.amount,
+        BigInt(contribution.resource_type),
+      );
+    });
+
     return tempPlayerPointsLeaderboard;
-  });
-
+  };
 
   // useEffect(() => {
   //   const playerOrder = getOrderName(
