+++ client/src/ui/components/HooksComponent.tsx
@@ -3,10 +3,12 @@ import { useFetchBlockchainData } from "../../hooks/store/useBlockchainStore";
 import { useComputeMarket } from "../../hooks/store/useMarketStore";
 import useUIStore from "@/hooks/store/useUIStore";
 import { Hexagon } from "@/types";
+import { useSetExistingStructures } from "@/hooks/store/_mapStore";
 
 export const HooksComponent = () => {
   useFetchBlockchainData();
   useComputeMarket();
+  useSetExistingStructures();
 
   const setHexData = useUIStore((state) => state.setHexData);
