+++ client/src/ui/components/models/buildings/worldmap/Structures.tsx
@@ -1,8 +1,5 @@
-import { useDojo } from "@/hooks/context/DojoContext";
-import { useEntityQuery } from "@dojoengine/react";
-import { Has, getComponentValue } from "@dojoengine/recs";
 import { useGLTF } from "@react-three/drei";
-import { useEffect, useMemo } from "react";
+import { useMemo } from "react";
 import * as THREE from "three";
 import { getUIPositionFromColRow } from "@/ui/utils/utils";
 import { StructureType } from "@bibliothecadao/eternum";
@@ -21,30 +18,7 @@ export const Structures = () => {
     [],
   );
 
-  const { setup } = useDojo();
-
   const existingStructures = useUIStore((state) => state.existingStructures);
-  const setExistingStructures = useUIStore((state) => state.setExistingStructures);
-
-  const builtStructures = useEntityQuery([Has(setup.components.Structure)]);
-
-  const _tmp = useMemo(() => {
-    return builtStructures.map((entity) => {
-      const position = getComponentValue(setup.components.Position, entity);
-      const structure = getComponentValue(setup.components.Structure, entity);
-      const type = StructureType[structure!.category as keyof typeof StructureType];
-      return {
-        col: position!.x,
-        row: position!.y,
-        type: type as StructureType,
-        entity: entity,
-      };
-    });
-  }, [builtStructures.length]);
-
-  useEffect(() => {
-    setExistingStructures(_tmp);
-  }, [_tmp]);
 
   return existingStructures.map((structure, index) => {
     return <BuiltStructure key={index} position={structure} models={models} structureCategory={structure.type} />;
